[fix] 유저 관계 매핑 (#155)
* fix: 유저 관계 설정 1차

* fix: 유저 관계 설정 2차

* be/fix: 유저 연관관계 매핑 - controller로 로직 이동

* be/fix: 리뷰 반영(converter로 로직 수정)

---------

+++ backend/src/main/java/com/techeer/backend/api/feedback/service/FeedbackService.java
@@ -4,12 +4,13 @@ import com.techeer.backend.api.aifeedback.domain.AIFeedback;
 import com.techeer.backend.api.aifeedback.repository.AIFeedbackRepository;
 import com.techeer.backend.api.feedback.converter.FeedbackConverter;
 import com.techeer.backend.api.feedback.domain.Feedback;
-import com.techeer.backend.api.feedback.dto.AllFeedbackResponse;
-import com.techeer.backend.api.feedback.dto.FeedbackCreateRequest;
-import com.techeer.backend.api.feedback.dto.FeedbackResponse;
+import com.techeer.backend.api.feedback.dto.request.FeedbackCreateRequest;
+import com.techeer.backend.api.feedback.dto.response.AllFeedbackResponse;
+import com.techeer.backend.api.feedback.dto.response.FeedbackResponse;
 import com.techeer.backend.api.feedback.repository.FeedbackRepository;
 import com.techeer.backend.api.resume.domain.Resume;
 import com.techeer.backend.api.resume.repository.ResumeRepository;
+import com.techeer.backend.api.user.domain.User;
 import com.techeer.backend.global.error.ErrorStatus;
 import com.techeer.backend.global.error.exception.BusinessException;
 import java.util.List;
@@ -27,18 +28,20 @@ public class FeedbackService {
     private final AIFeedbackRepository aiFeedbackRepository;
 
     @Transactional
-    public FeedbackResponse createFeedback(Long resumeId, FeedbackCreateRequest feedbackCreateRequest) {
+    public FeedbackResponse createFeedback(User user, Long resumeId, FeedbackCreateRequest feedbackCreateRequest) {
+
         Resume resume = resumeRepository.findByIdAndDeletedAtIsNull(resumeId)
                 .orElseThrow(() -> new BusinessException(ErrorStatus.RESUME_NOT_FOUND));
 
-        Feedback feedback = Feedback.of(resume, feedbackCreateRequest);
+        Feedback feedback = FeedbackConverter.toFeedbackEntity(user, resume, feedbackCreateRequest);
         feedbackRepository.save(feedback);
 
         return FeedbackConverter.toFeedbackResponse(feedback);
     }
 
     @Transactional
-    public void deleteFeedbackById(Long resumeId, Long feedbackId) {
+    public void deleteFeedbackById(User user, Long resumeId, Long feedbackId) {
+
         Resume resume = resumeRepository.findById(resumeId)
                 .orElseThrow(() -> new BusinessException(ErrorStatus.RESUME_NOT_FOUND));
 
@@ -49,21 +52,25 @@ public class FeedbackService {
             throw new BusinessException(ErrorStatus.INVALID_FEEDBACK_FOR_RESUME);
         }
 
+        if (!feedback.getUser().getId().equals(user.getId())) {
+            throw new BusinessException(ErrorStatus.UNAUTHORIZED);
+        }
+
         log.info("피드백 삭제 중: 피드백 ID {} (이력서 ID {})", feedbackId, resumeId);
 
         feedbackRepository.delete(feedback);
     }
 
-    @Transactional(readOnly = true)
     public AllFeedbackResponse getFeedbackWithAIFeedback(Long resumeId) {
+
         // resumeId에 해당하는 모든 피드백 가져오기
         List<Feedback> feedbacks = feedbackRepository.findAllByResumeId(resumeId);
         if (feedbacks.isEmpty()) {
             throw new BusinessException(ErrorStatus.FEEDBACK_NOT_FOUND);
         }
 
-        // resumeId에 해당하는 AI 피드백 가져오기 (없으면 null 설정)
-        AIFeedback aiFeedback = aiFeedbackRepository.findByResumeId(resumeId).orElse(null);
+        // resumeId에 해당하는 AI 피드백 가져오기 (없으면 빈 배열 반환)
+        AIFeedback aiFeedback = aiFeedbackRepository.findByResumeId(resumeId).orElse(AIFeedback.empty());
 
         // FeedbackConverter에서 모든 피드백 리스트와 AI 피드백을 포함한 응답으로 변환
         return FeedbackConverter.toAllFeedbackResponse(feedbacks, aiFeedback);
